+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- metrics-logstash-node.rb: Added queue_push_duration_in_millis metric for logstash pipeline inputs. (@Evesy)
 
 ## [1.1.1] - 2017-08-26
 ### Fixed